+++ packages/components/CHANGELOG.md
@@ -14,8 +14,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+#### Added
+
 - `Accordion`: Add experimental property `sticky` for a fixed heading
   ([#3181](https://github.com/porsche-design-system/porsche-design-system/pull/3181))
+- `Inline Notification`, `Banner`: heading hierarchy can now be customized with `headingTag` prop
+  ([#3168](https://github.com/porsche-design-system/porsche-design-system/pull/3168))
+
+#### Changed
+
+- `Accordion`: `tag` property is deprecated. Use `headingTag` property instead to specify heading hierarchy level.
+  ([#3168](https://github.com/porsche-design-system/porsche-design-system/pull/3168))
+
+```diff
+- <p-accordion tag="h3"></p-accordion>
++ <p-accordion heading-tag="h3"></p-accordion>
+```
+
+#### Fixed
+
+- `Pin Code`: Input is entered twice in iOS
+  ([#3192](https://github.com/porsche-design-system/porsche-design-system/pull/3192))
 
 ### [3.15.0-rc.1] - 2024-04-17